bookstore/bookstore-api/requirements.txt
fastapi[standard]==0.115.12
SQLAlchemy==2.0.40
python-dotenv==1.1.0
alembic==1.15.2
psycopg2-binary==2.9.10
bcrypt==4.3.0
PyJWT==2.10.1
pytest==8.3.5
httpx==0.28.1
